% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCor.R
\name{matrixCor}
\alias{matrixCor}
\title{Matrix Correlation}
\usage{
matrixCor(A)
}
\arguments{
\item{A}{The data matrix.}
}
\value{
The sample correlation matrix.
}
\description{
This function computes the sample correlation matrix from a data matrix using RcppArmadillo.
}
