% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixEigen.R
\name{matrixEigen}
\alias{matrixEigen}
\title{Eigenvalue Decomposition}
\usage{
matrixEigen(A)
}
\arguments{
\item{A}{A symmetric matrix.}
}
\value{
A list containing the eigenvalues and eigenvectors of matrix A.
}
\description{
This function computes the eigenvalue decomposition of a symmetric matrix using RcppArmadillo.
}
