% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixGeneralizedInverse.R
\name{matrixGeneralizedInverse}
\alias{matrixGeneralizedInverse}
\title{Generalized Inverse of a Matrix}
\usage{
matrixGeneralizedInverse(A, tol = 5e-16)
}
\arguments{
\item{A}{A numeric matrix.}

\item{tol}{Tolerance for singular values. Default is 1e-6.}
}
\value{
The generalized inverse of the input matrix.
}
\description{
This function computes the generalized inverse (Moore-Penrose pseudoinverse) of a matrix using RcppArmadillo.
}
