% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixInverse.R
\name{matrixInverse}
\alias{matrixInverse}
\title{Matrix Inversion}
\usage{
matrixInverse(A)
}
\arguments{
\item{A}{A square matrix.}
}
\value{
The inverse of matrix A.
}
\description{
This function computes the inverse of a square matrix using RcppArmadillo.
}
