% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixKronecker.R
\name{matrixKronecker}
\alias{matrixKronecker}
\title{Kronecker Product}
\usage{
matrixKronecker(A, B)
}
\arguments{
\item{A}{The first matrix.}

\item{B}{The second matrix.}
}
\value{
The Kronecker product of matrices A and B.
}
\description{
This function computes the Kronecker product of two matrices using RcppArmadillo.
}
