% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixListProduct.R
\name{matrixListProduct}
\alias{matrixListProduct}
\title{Product of a List of Matrices}
\usage{
matrixListProduct(matrixList)
}
\arguments{
\item{matrixList}{A list of numeric matrices.}
}
\value{
The product of the input matrices.
}
\description{
This function computes the product of a list of matrices in the order they appear using RcppArmadillo.
}
