% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixMultiply.R
\name{matrixMultiply}
\alias{matrixMultiply}
\title{Matrix Multiplication}
\usage{
matrixMultiply(A, B)
}
\arguments{
\item{A}{The first matrix.}

\item{B}{The second matrix.}
}
\value{
The product of matrices A and B.
}
\description{
This function performs matrix multiplication using RcppArmadillo.
}
