% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixSolve.R
\name{matrixSolve}
\alias{matrixSolve}
\title{Solve a Linear System}
\usage{
matrixSolve(A, B)
}
\arguments{
\item{A}{A square numeric matrix.}

\item{B}{A numeric vector or matrix.}
}
\value{
A matrix.
}
\description{
This function solves AX = B using RcppArmadillo.
}
