% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixSVD.R
\name{matrixSVD}
\alias{matrixSVD}
\title{Singular Value Decomposition}
\usage{
matrixSVD(A)
}
\arguments{
\item{A}{The matrix to decompose.}
}
\value{
A list containing d (singular values), u (left singular vectors), and v (right singular vectors).
}
\description{
This function performs singular value decomposition using RcppArmadillo.
}
