% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixSylvester.R
\name{matrixSylvester}
\alias{matrixSylvester}
\title{Solve Sylvester Equation using Armadillo}
\usage{
matrixSylvester(A, B, C)
}
\arguments{
\item{A}{A square matrix.}

\item{B}{A square matrix.}

\item{C}{A square matrix.}
}
\value{
The solution matrix X.
}
\description{
This function solves the Sylvester equation AX + XB = C using Armadillo's built-in solver.
}
