% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixVectorMultiply.R
\name{matrixVectorMultiply}
\alias{matrixVectorMultiply}
\title{Matrix-Vector Multiplication}
\usage{
matrixVectorMultiply(A, b)
}
\arguments{
\item{A}{The matrix.}

\item{b}{The vector.}
}
\value{
The product of matrix A and vector b.
}
\description{
This function performs matrix-vector multiplication using RcppArmadillo.
}
