% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixSylvester.R, R/matrixSylvesterEigen.R
\name{matrixSylvesterEigen}
\alias{matrixSylvesterEigen}
\title{Solve Sylvester Equation using Armadillo}
\usage{
matrixSylvesterEigen(UA, DA, UB, DB, C)

matrixSylvesterEigen(UA, DA, UB, DB, C)
}
\arguments{
\item{UA}{Eigenvector matrix of A.}

\item{DA}{Eigenvalues of A.}

\item{UB}{Eigenvector matrix of B.}

\item{DB}{Eigenvalues of B.}

\item{C}{A square matrix.}

\item{A}{A square matrix.}

\item{B}{A square matrix.}
}
\value{
The solution matrix X.

The solution matrix X.
}
\description{
This function solves the Sylvester equation AX + XB = C using Armadillo's built-in solver.

This function solves the Sylvester equation AX + XB = C using precomputed eigen decompositions of A and B.
}
